sanitizing dataflow code
