+++ rap/src/bin/rap.rs
@@ -49,7 +49,7 @@ fn main() {
             "-callgraph" => compiler.enable_callgraph(),
             "-mir" => compiler.enable_show_mir(),
             "-debug" => debug = true,
-            "-t" | "-test" => compiler.enable_test(),
+            "-D" | "-dataflow" => compiler.enable_dataflow(),
             _ => args.push(arg),
         }
     }
